WizardHeader and WizardContent
:: File Structure ::
Creates a subdirectory "WizardContent" inside of the Wizard dir. Moves
the WizardContent.tsx file to the "WizardContent" subdirectory. Updates
the import for WizardContent inside of the Wizard component.

:: WizardHeader ::
Creates three components in a file WizardHeader.tsx. The WizardProgress
displays a progress bar. StepTitle displays the title of the wizard and
the currentStep. WizardHeader renders out those two components.

:: WizardContent ::
Adds children to the list of props and renders them in the second
GridItem.